pyproject.toml
[project]
name = "auth"
version = "0.1.0"
description = "Lightweight authentication and authorization microservice"
readme = "README.md"
requires-python = ">=3.12"
license = { file = "LICENSE" }
authors = [
    {name = "Ananda Sarangaram"}
]
keywords = ["authentiation",  "JWT", ]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "alembic>=1.17.2",
    "basedpyright>=1.36.1",
    "bcrypt>=5.0.0",

    "fastapi>=0.127.0",
    "pydantic>=2.12.5",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.21",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.40.0",
]

[project.scripts]
auth-server = "auth.main:main"


[project.optional-dependencies]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
]
all = [
    "auth[dev]",
]
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/auth"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
python_files = ["test_*.py"]
asyncio_mode = "auto"
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
]
addopts = "--cov=auth --cov-report=html --cov-report=term-missing --cov-fail-under=90"

[tool.coverage.run]
source = ["src/auth"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/venv/*",
    "*/.venv/*",
    "*/typings/*",
    "*/main.py"  # CLI entry point, not tested
]

[tool.coverage.report]
precision = 2
skip_empty = true
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
    "@abc.abstractmethod",
]

[tool.ruff]
line-length = 100
target-version = "py312"
exclude = [
  "alembic/versions",
]

[tool.ruff.lint]
select = ["E", "F", "I", "W", "F811", "F401", "UP",
    "F841",  # local variable assigned but never used
    "B018",  # useless expression
  ]
ignore = ["E501"]
